Refactor App, remove filter componenet
